fix admin page error

+++ site/src/jsMain/kotlin/com/saitawngpha/blogmultiplatform/pages/admin/Create.kt
@@ -409,7 +409,7 @@ fun CreateScreen() {
                                         author = localStorage.getItem("username").toString(),
                                         title = uiState.title,
                                         subtitle = uiState.subtitle,
-                                        date = Date.now().toLong(),
+                                        date = Date.now(),
                                         thumbnail = uiState.thumbnail,
                                         content = uiState.content,
                                         category = uiState.category,
